✨ feat: Add ability to set custom buffer size for AudioRecord
The commit adds the ability to set a custom buffer size for AudioRecord by adding a new parameter to the plugin configuration. The default value for BUFFER_SIZE is now -1, indicating that it should be calculated based on the sample rate, channel configuration, and audio format. If a custom buffer size is specified in the configuration, it will be used instead of the calculated value. If the custom size is too small, the minimum required buffer size will be used instead.

Options:
- feat: Allow to set custom buffer size for AudioRecord
- fix: Set correct value for BUFFER_SIZE initialization
- chore: Refactor MicStreamPlugin.java for better readability
- docs: Update README file with new parameter description
- test: Add unit tests for custom buffer size functionality